Revert "rename unit tests to end in LayoutRendererTests"
This reverts commit 276705e5d2ee605734493ad840f952aa123d98a6.
